Remove PipelineRun delete buttons
